changes to get socket transceiver working
